Add configurable processing fee support for fiat payments (#51)
* Initial plan

* Add processing fee support for fiat payments (Revolut 1% + 0.20 EUR)


* Add tests for processing fee calculation


* Improve processing fee comments and clarify TODOs


* Make processing fees configurable via settings


* Fix processing fee calculation rounding and add currency parse error logging


* Use settings types directly instead of duplicate structs


* Add logging for invalid currency config and clarify default comment


* feat: database payment method config

* feat: require company_id

* chore: [dev] insert company

* fix: store revolut webhook

* fix: convert processing fee f32

* fix: hide secrets

* chore: disable bitvora

* fix: partial payment config updates

* refactor: change all admin endpoints to use u64 money values

* fix: migrate cost plan amounts

---------

+++ lnvps_api_common/src/model.rs
@@ -406,7 +406,8 @@ pub struct ApiVmCostPlan {
     pub id: u64,
     pub name: String,
     pub currency: ApiCurrency,
-    pub amount: f32,
+    /// Cost amount in smallest currency units (cents for fiat, millisats for BTC)
+    pub amount: u64,
     pub other_price: Vec<ApiPrice>,
     pub interval_amount: u64,
     pub interval_type: ApiVmCostPlanIntervalType,
